issue/317 - Rename current llama folder to llama_legacy
