+++ csrc/models/llama_legacy/llama_decoder_layer.cpp
@@ -3,7 +3,7 @@
 #include "infinicore/ops.hpp"
 #include <optional>
 
-namespace infinilm::models::llama {
+namespace infinilm::models::llama_legacy {
 /**
  * @deprecated This function is deprecated and will be REMOVED in the next major release (v0.2.0).
  *
@@ -78,4 +78,4 @@ LlamaDecoderLayer::forward(infinicore::Tensor &hidden_states,
     return std::make_tuple(hidden_states, residual);
 }
 
-} // namespace infinilm::models::llama
+} // namespace infinilm::models::llama_legacy
