+++ csrc/models/llama_legacy/llama_model.cpp
@@ -5,7 +5,7 @@
 #include "infinicore/ops.hpp"
 #include <iostream>
 
-namespace infinilm::models::llama {
+namespace infinilm::models::llama_legacy {
 /**
  * @deprecated This function is deprecated and will be REMOVED in the next major release (v0.2.0).
  *
@@ -176,4 +176,4 @@ void LlamaModel::reset_cache(const cache::CacheConfig *cache_config) {
     }
 }
 
-} // namespace infinilm::models::llama
+} // namespace infinilm::models::llama_legacy
