+++ csrc/models/model_factory.cpp
@@ -1,5 +1,5 @@
 #include "model_factory.hpp"
-#include "llama/llama_for_causal_lm.hpp"
+#include "llama_legacy/llama_for_causal_lm.hpp"
 #include "models_registry.hpp"
 
 namespace infinilm {
@@ -21,9 +21,9 @@ std::shared_ptr<InfinilmModel> InfinilmModelFactory::createModel(
     const cache::CacheConfig *cache,
     backends::AttentionBackend attention_backend) {
     std::shared_ptr<InfinilmModel> model;
-    if (const auto llama_config_ptr = dynamic_cast<const models::llama::LlamaConfig *>(&config)) {
+    if (const auto llama_config_ptr = dynamic_cast<const models::llama_legacy::LlamaConfig *>(&config)) {
         const auto &llama_config = *llama_config_ptr;
-        model = std::make_shared<models::llama::LlamaForCausalLM>(
+        model = std::make_shared<models::llama_legacy::LlamaForCausalLM>(
             llama_config, rank_info.device, rank_info, attention_backend);
     } else {
         throw std::invalid_argument("InfinilmModelFactory::createModel: Unsupported model config type");
@@ -43,7 +43,7 @@ std::shared_ptr<InfinilmModel> InfinilmModelFactory::createModel(
     backends::AttentionBackend attention_backend) {
     std::shared_ptr<InfinilmModel> model;
     if (true) {
-        model = std::make_shared<models::llama::LlamaForCausalLM>(
+        model = std::make_shared<models::llama_legacy::LlamaForCausalLM>(
             model_config, rank_info.device, rank_info, attention_backend);
     } else {
         throw std::invalid_argument("InfinilmModelFactory::createModel: Unsupported model config type");
