+++ csrc/pybind11/bindings.cc
@@ -2,7 +2,7 @@
 
 #include "cache/cache.hpp"
 #include "engine/engine.hpp"
-#include "models/llama.hpp"
+#include "models/llama_legacy.hpp"
 
 namespace py = pybind11;
 
@@ -10,7 +10,7 @@ PYBIND11_MODULE(_infinilm, m) {
     m.doc() = "InfiniLM Llama model Python bindings";
 
     infinilm::cache::bind_cache(m);
-    infinilm::models::llama::bind_llama(m);
+    infinilm::models::llama_legacy::bind_llama(m);
     infinilm::engine::distributed::bind_dist_config(m);
     infinilm::engine::bind_infer_engine(m);
 }
